prefer / 신호등 저장 기능 및 값 수정, 마이트래픽 오더넘버 추가

+++ src/main/java/com/example/project/entity/MyTrafficEntity.java
@@ -32,7 +32,7 @@ public class MyTrafficEntity {
     private MemberEntity memberEntity;
 
     @Column(name = "myTrafficOrderNumber")
-    private Double myTrafficOrderNumber;
+    private Long myTrafficOrderNumber;
 
     @OneToMany(mappedBy = "myTrafficEntity", cascade = CascadeType.ALL, orphanRemoval = true, fetch = FetchType.LAZY)
     private List<TrafficTimeEntity> trafficTimeEntityList = new ArrayList<>();
